Address review comments
- extract wrapper from helpers.go to a new logger.go
- extract max buffer size constant
- log lines with no specified level at INFO
